+++ modules/test/poshi-runner/poshi-runner/src/main/java/com/liferay/poshi/runner/prose/PoshiProseScenario.java
@@ -33,6 +33,8 @@ import org.dom4j.tree.DefaultAttribute;
 public class PoshiProseScenario extends BasePoshiProse {
 
 	public PoshiProseScenario(String poshiProseScenarioString) {
+		poshiProseScenarioString = filterCommentLines(poshiProseScenarioString);
+
 		Matcher setupMatcher = _setupPattern.matcher(poshiProseScenarioString);
 
 		String tags = null;
@@ -84,10 +86,6 @@ public class PoshiProseScenario extends BasePoshiProse {
 			_scenarioContent, PoshiProseStatement.KEYWORDS);
 
 		for (String poshiProseStatementString : poshiProseStatementStrings) {
-			if (poshiProseStatementString.startsWith("#")) {
-				continue;
-			}
-
 			_poshiProseStatements.add(
 				new PoshiProseStatement(poshiProseStatementString));
 		}